Prepare to add commands
