+++ macros/src/lib.rs
@@ -6,11 +6,11 @@ mod from_file;
 mod from_macro;
 
 #[proc_macro_attribute]
-pub fn implement_registers_from_file(attr: TokenStream, item: TokenStream) -> TokenStream {
-    from_file::implement_registers_from_file(attr, item)
+pub fn implement_device_from_file(attr: TokenStream, item: TokenStream) -> TokenStream {
+    from_file::implement_device_from_file(attr, item)
 }
 
 #[proc_macro]
-pub fn implement_registers(item: TokenStream) -> TokenStream {
-    from_macro::implement_registers(item.into()).into()
+pub fn implement_device(item: TokenStream) -> TokenStream {
+    from_macro::implement_device(item.into()).into()
 }
