+++ test-files/yaml_syntax.yaml
@@ -1,4 +1,4 @@
-address_type: u8
+register_address_type: u8
 registers:
   id:
     description: Doc comment for the ID register
